requirements.txt
pystray
pynput
Pillow
requests
win11toast
pyperclip
beautifulsoup4
pywin32
